Reorganize after-form-submit routines
