+++ src/app/pages/adoption/adoption.js
@@ -5,8 +5,13 @@ import RoleSelector from '~/components/role-selector/role-selector';
 import StudentForm from '~/components/student-form/student-form';
 import MultiPageForm from '~/components/multi-page-form/multi-page-form';
 import ContactInfo from '~/components/contact-info/contact-info';
-import BookSelector, {useSelectedBooks, useAfterSubmit, useFirstSearchArgument}
-    from '~/components/book-selector/book-selector';
+import {
+    useAfterSubmit,
+    useFirstSearchArgument
+} from '~/components/book-selector/after-form-submit';
+import BookSelector, {
+    useSelectedBooks
+} from '~/components/book-selector/book-selector';
 import HowUsing from './how-using/how-using';
 import useSalesforceContext from '~/contexts/salesforce';
 import useFormTarget from '~/components/form-target/form-target';
